+++ packages/metro-service/src/assets-registry/registry.js
@@ -13,17 +13,33 @@
 
 'use strict';
 
+/*::
+export type AssetDestPathResolver = 'android' | 'generic';
 
+export type PackagerAsset = {
+  +__packager_asset: boolean,
+  +fileSystemLocation: string,
+  +httpServerLocation: string,
+  +width: ?number,
+  +height: ?number,
+  +scales: Array<number>,
+  +hash: string,
+  +name: string,
+  +type: string,
+  +resolver?: AssetDestPathResolver,
+  ...
+};
+*/
 
-const assets = [];
+const assets /*: Array<PackagerAsset> */ = [];
 
-function registerAsset(asset) {
+function registerAsset(asset /*: PackagerAsset */) /*: number */ {
   // `push` returns new array length, so the first asset will
   // get id 1 (not 0) to make the value truthy
   return assets.push(asset);
 }
 
-function getAssetByID(assetId) {
+function getAssetByID(assetId /*: number */) /*: PackagerAsset */ {
   return assets[assetId - 1];
 }